Add quadlet (#119)
Add quadlet binary and register it as systemd unit generator.
When the tar archive is installed on a systemd-enabled host, the quadlet is called by systemd on `systemctl daemon-reload` to generate unit files based on *.container files within `/etc/containers/systemd`.

The quadlet binary is not contained within the minimal image (because it is actually meant to be published with the tar archive only since systemd is not used inside the container).